[Refactor] put typeDef in a separate file
