+++ src/schema.js
@@ -1,0 +1,16 @@
+const { gql } = require('apollo-server-express')
+
+module.exports = gql`
+    type Query {
+        notes: [Note]!
+        note(id: ID!): Note!
+    },
+    type Note {
+        id: ID!
+        content: String!
+        author: String!
+    },
+    type Mutation {
+        newNote(content: String!, author: String!): Note!
+    }
+`
